parsers/banki_ru_parser/requirements.txt
requests==2.31.0
kafka-python==2.0.2
redis==4.5.5
